Remove all assembly for popcnt
On some systems, running `goapp test` fails to set the "appengine" build
flag. That causes it to try compiling the assembly version.  Of course
that doesn't work because App Engine doesn't support assembly.

This is an unfortunate work around for specific systems under which I
was unable to make `gapp test` work correctly.  This commit should not
be sent upstream.